+++ src/main/java/com/johnsproject/jpge2/dto/Transform.java
@@ -28,9 +28,9 @@ import com.johnsproject.jpge2.processors.VectorProcessor;
 
 public class Transform {
 
-	private static final int vx = VectorProcessor.VECTOR_X;
-	private static final int vy = VectorProcessor.VECTOR_Y;
-	private static final int vz = VectorProcessor.VECTOR_Z;
+	private static final byte vx = VectorProcessor.VECTOR_X;
+	private static final byte vy = VectorProcessor.VECTOR_Y;
+	private static final byte vz = VectorProcessor.VECTOR_Z;
 
 	private int[] location;
 	private int[] rotation;
