+++ src/main/java/com/johnsproject/jpge2/shaders/GouraudSpecularShader.java
@@ -19,9 +19,9 @@ import com.johnsproject.jpge2.processors.GraphicsProcessor.Shader;
 
 public class GouraudSpecularShader implements Shader {
 
-	private final int vx = VectorProcessor.VECTOR_X;
-	private final int vy = VectorProcessor.VECTOR_Y;
-	private final int vz = VectorProcessor.VECTOR_Z;
+	private final byte vx = VectorProcessor.VECTOR_X;
+	private final byte vy = VectorProcessor.VECTOR_Y;
+	private final byte vz = VectorProcessor.VECTOR_Z;
 
 	private final int[] uvX = VectorProcessor.generate();
 	private final int[] uvY = VectorProcessor.generate();
@@ -106,7 +106,7 @@ public class GouraudSpecularShader implements Shader {
 				break;
 			case POINT:
 				// attenuation
-				int distance = VectorProcessor.magnitude(lightDirection);
+				long distance = VectorProcessor.magnitude(lightDirection);
 				int attenuation = MathProcessor.FP_VALUE;
 				attenuation += MathProcessor.multiply(distance, 3000);
 				attenuation += MathProcessor.multiply(MathProcessor.multiply(distance, distance), 20);
